Add folders for stories by group
