+++ stories/authentication/loginButton.stories.tsx
@@ -21,8 +21,8 @@
 
 import React, { ReactElement } from "react";
 import { Button } from "@material-ui/core";
-import { SessionProvider } from "../src/context/sessionContext";
-import { LoginButton } from "../src/components/logIn";
+import { SessionProvider } from "../../src/context/sessionContext";
+import { LoginButton } from "../../src/components/logIn";
 
 export default {
   title: "Authentication/Login Button",
