+++ stories/components/text.stories.tsx
@@ -21,10 +21,10 @@
 
 import React, { ReactElement } from "react";
 import * as SolidFns from "@inrupt/solid-client";
-import { Text } from "../src/components/text";
-import { DatasetProvider } from "../src/context/datasetContext";
-import { ThingProvider } from "../src/context/thingContext";
-import config from "./config";
+import { Text } from "../../src/components/text";
+import { DatasetProvider } from "../../src/context/datasetContext";
+import { ThingProvider } from "../../src/context/thingContext";
+import config from "../config";
 
 const { host } = config();
