+++ stories/components/video.stories.tsx
@@ -21,7 +21,7 @@
 
 import React, { ReactElement } from "react";
 import * as SolidFns from "@inrupt/solid-client";
-import { Video } from "../src/components/video";
+import { Video } from "../../src/components/video";
 
 export default {
   title: "Components/Video",
